requirements.txt
PyQt5>=5.15.0
python-xlib>=0.33
